Alterando o tipo da propriedade Id de Int para long em PaymentDTO e ProductDTO

+++ Invoice.Services/DTO/ItemDTO.cs
@@ -8,7 +8,7 @@ namespace InvoiceApi.Services.DTO
 {
     public class ItemDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Description { get; set; }
         public decimal UnitValue { get; set; }
         public int Quantity { get; set; }
@@ -19,7 +19,7 @@ namespace InvoiceApi.Services.DTO
         public ItemDTO()
         {}
 
-        public ItemDTO(int id, string description, decimal unitValue, int quantity, decimal totalItem, long invoiceId, long productId)
+        public ItemDTO(long id, string description, decimal unitValue, int quantity, decimal totalItem, long invoiceId, long productId)
         {
             Id = id;
             Description = description;
